backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# Database
psycopg2-binary
alembic
# Utilities
httpx
email-validator
